// src/database/generated/models/UserSession.ts
export type UserSessionModel = runtime.Types.Result.DefaultSelection<Prisma.$UserSessionPayload>
// ...
export type $UserSessionPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "UserSession"
  objects: {
    user: Prisma.$UserPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    userId: number
    refreshToken: string
    userAgent: string | null
    ipAddress: string | null
    isActive: boolean
    expiresAt: Date
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["userSession"]>
  composites: {}
}